Moving the GraphicsMemory* types to the TerraFX.Advanced namespace
