+++ sources/Graphics/Core/Graphics/GraphicsTexture.cs
@@ -1,6 +1,7 @@
 // Copyright © Tanner Gooding and Contributors. Licensed under the MIT License (MIT). See License.md in the repository root for more information.
 
 using System;
+using TerraFX.Advanced;
 
 namespace TerraFX.Graphics;
